Refactor 1: use project references for compiling common lib and server code separately.
