+++ common/src/game-setup-types.ts
