+++ socket/src/model.d.ts
